move get_external_market_center_price from base to relative orders strategy
